Navigate to Set password after a success
